services/service-projection/requirements.txt
# Projection Service Requirements

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async and networking
httpx==0.25.2
aiofiles==23.2.1

# Database connections
clickhouse-driver==0.2.6
redis==5.0.1

# Kafka integration
aiokafka==0.8.1

# Monitoring and observability
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
structlog==23.2.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# Configuration and utilities
pyyaml==6.0.1
python-dotenv==1.0.0
tenacity==8.2.3

# Testing (for development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2  # For testing

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
